Rename module integrated to module plugin, establish new documentation
