+++ .idea/encodings.xml
@@ -2,8 +2,6 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/tickbox-api/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/tickbox-integrated-tests/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/tickbox-integrated-tests/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/tickbox-integrated/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/tickbox-plugin/src/main/java" charset="UTF-8" />
   </component>
 </project>
