+++ .idea/jarRepositories.xml
@@ -2,9 +2,19 @@
 <project version="4">
   <component name="RemoteRepositoriesConfiguration">
     <remote-repository>
+      <option name="id" value="snapshots-repo" />
+      <option name="name" value="snapshots-repo" />
+      <option name="url" value="https://repo.auriium.xyz/snapshots" />
+    </remote-repository>
+    <remote-repository>
       <option name="id" value="central" />
       <option name="name" value="Central Repository" />
       <option name="url" value="https://repo.maven.apache.org/maven2" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="central-repo" />
+      <option name="name" value="central-repo" />
+      <option name="url" value="https://repo.auriium.xyz/releases" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="central-repo" />
