+++ tickbox-api/src/main/java/xyz/auriium/tick/Tick.java
@@ -1,12 +1,13 @@
 package xyz.auriium.tick;
 
-import xyz.auriium.tick.container.ContainerManager;
+import xyz.auriium.tick.container.container.TickContainer;
+import xyz.auriium.tick.container.terms.CreationTerms;
 
 /**
  * Represents a prestarted Tick that can provide multiple containers
  */
 public interface Tick {
 
-    ContainerManager getManager();
+    <T extends TickContainer> T startupContainer(CreationTerms<T> terms) throws InterruptedException;
 
 }
