+++ tickbox-api/src/main/java/xyz/auriium/tick/docker/source/impl/SystemEnvSourceProvider.java
@@ -4,6 +4,11 @@ import com.github.dockerjava.core.DefaultDockerClientConfig;
 import com.github.dockerjava.core.DockerClientConfig;
 import xyz.auriium.tick.docker.source.*;
 
+/**
+ * Provider that attempts to use system variables in order to provide a docker source
+ *
+ * Analogous to TestContainer's EnvironmentAndSystemPropertyClientProviderStrategy (geez)
+ */
 public class SystemEnvSourceProvider implements DockerSourceProvider {
     @Override
     public String name() {
@@ -25,7 +30,7 @@ public class SystemEnvSourceProvider implements DockerSourceProvider {
 
     @Override
     public ApplicableResult isApplicable() {
-        return null;
+        return System.getenv("DOCKER_HOST") != null ? ApplicableResult.success() : ApplicableResult.fail("System Environment Variables for docker could not be located or are null! Please fill them out!");
     }
