--- tickbox-api/src/test/java/xyz/auriium/tick/DockerStartupTest.java
@@ -1,4 +1,0 @@
-package xyz.auriium.tick;
-
-public class DockerStartupTest {
-}
