+++ tickbox-api/src/test/java/xyz/auriium/tick/StartupIT.java
@@ -1,0 +1,18 @@
+package xyz.auriium.tick;
+
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Test that tests startup and initialization for each existing strategy
+ */
+public class StartupIT {
+
+    @Test
+    public void initializeTick() {
+        Tick tick = new CommonTickFactory(null,null,null,null).produce();
+    }
+
+
+
+}
