+++ tickbox-plugin/src/main/java/xyz/auriium/tick/plugin/AbstractTickMojo.java
@@ -1,4 +1,4 @@
-package xyz.auriium.tick.integrated;
+package xyz.auriium.tick.plugin;
 
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugins.annotations.Parameter;
