+++ tickbox-plugin/src/main/java/xyz/auriium/tick/plugin/CommonInits.java
@@ -1,4 +1,4 @@
-package xyz.auriium.tick.integrated;
+package xyz.auriium.tick.plugin;
 
 public enum CommonInits {
