+++ tickbox-plugin/src/main/java/xyz/auriium/tick/plugin/Initializer.java
@@ -1,4 +1,4 @@
-package xyz.auriium.tick.integrated;
+package xyz.auriium.tick.plugin;
 
 import xyz.auriium.tick.container.terms.JDBCTerms;
